requirements.txt
# ── LangChain core ───────────────────────────────────────────────────────────
langchain==0.3.25
# langchain-core==0.3.56
langchain-core>=0.3.58
langchain-openai==0.3.16

# ── LLM providers (pick one or both) ─────────────────────────────────────────
openai==1.72.0
# langchain-google-genai==2.1.4   # uncomment if using Gemini

# ── Data validation & serialization ──────────────────────────────────────────
pydantic==2.11.3

# ── EPO OPS API client (replaces Lens.org) ───────────────────────────────────
python-epo-ops-client==4.2.1
xmltodict==0.14.2                 # fallback XML parser for endpoints that don't support JSON

# ── Data processing & output ──────────────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5                   # pandas Excel export (.xlsx)

# ── Local caching (SQLite) ────────────────────────────────────────────────────
diskcache==5.6.3                  # cache EPO API responses to disk

# ── Environment & config ──────────────────────────────────────────────────────
python-dotenv==1.1.0

# ── Optional: PubMed support ──────────────────────────────────────────────────
# biopython==1.85                 # uncomment to enable PubMed fetcher

# ── Optional: async batch processing ─────────────────────────────────────────
# aiohttp==3.11.18                # uncomment for async patent fetching

pdfplumber